requirements.txt
Django==5.2.1
djangorestframework==3.15.0
django-allauth==0.61.0
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-filter==23.5
Pillow==10.1.0
stripe==7.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
django-tenants==3.5.0
psycopg2-binary==2.9.9
pymongo==3.11.4
setuptools>=60.0.0 